+++ package.json
@@ -2,17 +2,16 @@
   "name": "employee-manager",
   "version": "1.0.0",
   "description": "",
-  "main": "server.js",
+  "main": "index.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "node server.js"
+    "start": "node index.js"
   },
   "keywords": [],
   "author": "",
   "license": "ISC",
   "dependencies": {
     "console.table": "^0.10.0",
-    "express": "^4.18.2",
     "inquirer": "^8.2.4",
     "mysql2": "^3.1.0"
   }
