--- routes/api/deptRoutes.js
@@ -1,2 +1,0 @@
-const router = require('express').Router();
-const Book = require('../../models/Book');
