--- routes/index.js
@@ -1,6 +1,0 @@
-const router = require('express').Router();
-const apiRoutes = require('./api');
-
-router.use('/api', apiRoutes);
-
-module.exports = router;
